Update rooms processing
If a user passing a comma separated HUBOT_HIPCHAT_ROOMS list included a space in the room names (like many people probably do), then the system fails when trying to join the second room because the space is being included in the name.  This ensures that the jid it tries to use is using the "trimmed" version with beginning and ending whitespace removed.